drop prerender concurrency to 1
Since nuxt 3.8.0, static generation takes a lot of memory and ends up
failing with a "JavaScript heap out of memory" error most of the time on
our OpenShift instance.
One of the changes in the last versions of Nuxt that might cause this is the increase of the
nitro prerender concurrency to 4.