start: reduce layout shift
Because this page has both a static and a dynamic version, I wasn't able
to add a loading indicator. (The static version of the site seems to
overrule any sort of animation that would be displayed while the dynamic
content is being fetched.)

I was, however, able to add a little more content to the static version
of the site so that things don't jump around quite as much when the
dynamic content arrives.