🌐 locales: Correct Arabic language picker selection
Per discussion in `#l10n:fedoraproject.org` on Matrix, this string is
incorrectly translated for Arabic. Instead of translating to "Arabic",
the string translates to "Arab" which is mildly insensitive. This commit
corrects the mistake, and changes the translation to "Arabic" in Arabic.

Thanks `@basses:matrix.org` for pointing out this mistake on Matrix.

Signed-off-by: Justin W. Wheeler <jwf@redhat.com>